requirements.txt
opencv-python-headless>=4.5.1
numpy>=1.21.0
torch>=1.9.0
facenet-pytorch>=2.5.2
Pillow>=9.0.0